Used cfg(test) on some functions over allow dead code
This way we do not need any comments and are more disruptive with code.
It allows reduces compile time for the user of the library

I removed some previously annotated function with allow dead code
Reason: Even the tests do not use them at all.
They were private to the library anyway, so the user only would get
more compile time.
Dead code can be retrieved via source code control if needed later

2. Fixed clippy warning about auto-deref

3. Removed unused code for ThreadSafe trait:
ThreadSafe is not used anywhere.
Might be a relict of the past ?